backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
stripe==11.4.1
aiohttp==3.11.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==5.2.1
tenacity==9.0.0
langgraph==0.2.61
langchain-openai==0.3.0
langchain-core==0.3.29
langsmith==0.3.3
inngest==0.4.17
prometheus-client==0.21.1
httpx==0.28.1
xero-python==3.0.0
pyjwt[crypto]==2.10.1
python-dotenv==1.0.1
